Rename Preact version-specific modules to include major version
Rename these modules to make it clearer which version(s) of Preact they
deal with.
