+++ src/preact10-internals.ts
